requirements.txt
Flask==3.1.2
numpy==1.26.4
Pillow==11.3.0
python-dotenv==1.2.1
tensorflow==2.20.0
